backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
yt_dlp
python-dotenv
